requirements.txt
astroid==1.6.5
colorama==0.3.9
dj-database-url==0.5.0
Django==1.11.4
django-appconf==1.0.2
django-imagekit==4.0.2
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pilkit==2.0
Pillow==5.0.0
pylint==1.9.2
pytz==2018.5
six==1.11.0
virtualenv==16.0.0
whitenoise==4.1
wrapt==1.10.11
psycopg2==2.5.4